Stabilize Rest live stats
